refact : 외부 api를 api 디렉토리로 이동하였습니다.
