+++ src/main/resources/LAWD_CD.txt
@@ -1,0 +1,461 @@
+11000
+11110
+11140
+11170
+11200
+11215
+11230
+11260
+11290
+11305
+11320
+11350
+11380
+11410
+11440
+11470
+11500
+11530
+11545
+11560
+11590
+11620
+11650
+11680
+11710
+11740
+21000
+21110
+21140
+21170
+21200
+21230
+21260
+21290
+21320
+21350
+21380
+21410
+21440
+22000
+22110
+22140
+22170
+22200
+22230
+22260
+22290
+23000
+23110
+23140
+23170
+23200
+23230
+23260
+24000
+24110
+24140
+24170
+24200
+25000
+25110
+25140
+25170
+25200
+25230
+26000
+26110
+26140
+26170
+26200
+26230
+26260
+26290
+26320
+26350
+26380
+26410
+26440
+26470
+26500
+26530
+26710
+27000
+27110
+27140
+27170
+27200
+27230
+27260
+27290
+27710
+28000
+28110
+28140
+28170
+28177
+28185
+28200
+28230
+28237
+28245
+28260
+28710
+28720
+29000
+29110
+29140
+29155
+29170
+29200
+30000
+30110
+30140
+30170
+30200
+30230
+31000
+31110
+31140
+31170
+31200
+31710
+36110
+41000
+41010
+41011
+41012
+41013
+41014
+41050
+41110
+41111
+41113
+41115
+41117
+41130
+41131
+41133
+41135
+41150
+41170
+41171
+41173
+41190
+41191
+41193
+41195
+41197
+41199
+41210
+41220
+41230
+41250
+41270
+41271
+41273
+41280
+41281
+41283
+41285
+41287
+41290
+41310
+41330
+41350
+41360
+41370
+41390
+41410
+41430
+41450
+41460
+41461
+41463
+41465
+41480
+41500
+41550
+41570
+41590
+41610
+41630
+41650
+41670
+41710
+41720
+41730
+41740
+41750
+41760
+41770
+41780
+41790
+41800
+41810
+41820
+41830
+41840
+41850
+41860
+41870
+41880
+41890
+42000
+42050
+42110
+42130
+42150
+42170
+42190
+42210
+42230
+42710
+42715
+42720
+42730
+42740
+42745
+42750
+42760
+42770
+42780
+42790
+42800
+42810
+42820
+42830
+42840
+42850
+43000
+43050
+43110
+43111
+43112
+43113
+43114
+43130
+43150
+43710
+43720
+43730
+43740
+43745
+43750
+43760
+43770
+43780
+43790
+43795
+43800
+44000
+44050
+44110
+44111
+44113
+44115
+44130
+44131
+44133
+44150
+44170
+44180
+44190
+44200
+44210
+44230
+44250
+44270
+44710
+44720
+44730
+44740
+44750
+44760
+44770
+44780
+44790
+44800
+44810
+44820
+44825
+44830
+44840
+44850
+44860
+45000
+45050
+45110
+45111
+45113
+45130
+45140
+45150
+45170
+45180
+45190
+45210
+45710
+45720
+45730
+45740
+45750
+45760
+45770
+45780
+45790
+45800
+45810
+45820
+45830
+46000
+46010
+46011
+46012
+46013
+46020
+46030
+46050
+46110
+46130
+46150
+46170
+46190
+46210
+46230
+46710
+46720
+46730
+46740
+46750
+46760
+46770
+46780
+46790
+46800
+46810
+46820
+46830
+46840
+46850
+46860
+46870
+46880
+46890
+46900
+46910
+47000
+47010
+47011
+47012
+47013
+47014
+47015
+47016
+47020
+47050
+47110
+47111
+47113
+47130
+47150
+47170
+47190
+47210
+47230
+47250
+47270
+47280
+47290
+47710
+47720
+47730
+47740
+47750
+47760
+47770
+47780
+47790
+47795
+47800
+47810
+47820
+47830
+47840
+47850
+47860
+47870
+47880
+47890
+47900
+47910
+47920
+47930
+47940
+48000
+48010
+48011
+48012
+48013
+48014
+48015
+48016
+48050
+48060
+48070
+48110
+48120
+48121
+48123
+48125
+48127
+48129
+48130
+48131
+48133
+48135
+48137
+48150
+48151
+48153
+48160
+48170
+48190
+48210
+48220
+48230
+48240
+48250
+48270
+48290
+48310
+48330
+48710
+48720
+48730
+48740
+48750
+48760
+48770
+48775
+48780
+48790
+48795
+48800
+48810
+48820
+48830
+48840
+48850
+48860
+48870
+48880
+48890
+49000
+49110
+49130
+49710
+49720
+50000
+50110
+50130
