+++ src/test/kotlin/estert/domain/kakao/KakaoConfigTest.kt
@@ -1,16 +1,17 @@
 package estert.domain.kakao
 
+import estert.domain.api.kakao.config.KakaoProperties
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldNotBe
 
 class KakaoConfigTest : StringSpec({
         "Kakao Yaml 불러오기" {
             val yaml = org.yaml.snakeyaml.Yaml(
-                org.yaml.snakeyaml.constructor.Constructor(estert.domain.kakao.config.KakaoProperties::class.java))
+                org.yaml.snakeyaml.constructor.Constructor(KakaoProperties::class.java))
             // resources\kakao.yml
             val inputStream = this.javaClass.classLoader
                 .getResourceAsStream("kakao.yml")
-            val kakaoProperties = yaml.load<estert.domain.kakao.config.KakaoProperties>(inputStream)
+            val kakaoProperties = yaml.load<KakaoProperties>(inputStream)
 
             println(kakaoProperties)
             println(kakaoProperties.kakao)
